Chanakya/agent/pyproject.toml
[tool.black]
target-version = ["py313"]

[project]
name = "chanakya"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.3.1",
    "curl-cffi>=0.15.0",
    "edge-tts>=7.2.8",
    "isort>=8.0.1",
    "langchain>=1.2.15",
    "langchain-google-genai>=1.2.1",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
    "trafilatura>=2.0.0",
    "uvicorn>=0.32.0",
    "fastapi[standard]>=0.115.0",
]
